Deal with authorship properly, in a future-sustainable fashion
This module is arguably one of the most successful collaborative projects on
CPAN. Make it known to the world by introducing a clear canonical source of
authorship metadata.

I have been holding off jumping on the "contributors from git" bandwagon
because of multiple issues

- Not every contribution results in a direct, attributtable commit
- We already had an extensive contributor list, much larger than the shortlog
- Loss of ircnicks (a fine tradition started in 266bdcc3)

So this commit does the proverbial bullet-biting and introduces *proper*
handling of all the ins and outs, including author-mode (CI as well) tests
for proper attribution. As a side-effect we break a lot of the footer links,
to be fixed in next commit.

Btw as I write this `195 contributors made this library what it is today`...
Mind. Blown.