Clarify licensing, ensure footers are consistent throughout the project
Note - this *DOES NOT CHANGE* the licensing of the project in any way, it
is still dual-licensed like perl5, just as it was before this commit. All
additions are for clarifying purposes only.

The way licensing is presented is an amalgamation of information and
prior art found at the following sources:
 - http://www.perlfoundation.org/cpan_licensing_guidelines
 - https://metacpan.org/source/ETHER/Moose-2.1212/LICENSE
 - `git log --color -w -p v5.20.0 ^perl-5.000^ -- Artistic`