Unbreak $rs->create() with empty hashref on Oracle
Oracle does not support the INSERT INTO ... DEFAULT VALUES syntax used by
SQLMaker by default. Furthermore, Oracle has no way of inserting a row
without specifying any columns, so the only reasonably fix seems to be to
pick a column and supply the DEFAULT keyword as value for it.

This approach seems to work without problems, even for things like
sequence+trigger-as-auto-increment.