add ResultSet::rs to force scalar context
rs is a simple helper method that does nothing but return the original
ResultSet in order to force scalar context for search methods.  This is
primarily intended for predefined queries in subclasses to allow them to
take advantage of search's context-aware return values without requiring
them to either manually force scalar context or define a second XXX_rs
search method to get a resultset object in list context.