Take a copy instead of weakening in 5.8 leak workaround
I can't see a way $self could be the only strong reference to the
object: either $_[0] or some other copy would have to be strong in order
to have something to call the method on in the first place.  However,
let's not have everyone reading or editing the code have to do that
reasoning.