More desc_env fixes: Win32 lib matching, and better formatting of final diag
This is what the slash-situation is on a typical strawberry:

          'inc' => [
                     'C:/Strawberry/perl/site/lib/MSWin32-x64-multi-thread',
                     'C:/Strawberry/perl/site/lib',
                     'C:/Strawberry/perl/vendor/lib',
                     'C:/Strawberry/perl/lib',
                     '.'
                   ],
          'archlib' => 'C:\\STRAWB~1\\perl\\lib',
          'archlibexp' => 'C:\\STRAWB~1\\perl\\lib',
          'privlib' => 'C:\\STRAWB~1\\perl\\lib',
          'privlibexp' => 'C:\\STRAWB~1\\perl\\lib',
          'sitearch' => 'C:\\STRAWB~1\\perl\\site\\lib',
          'sitearchexp' => 'C:\\STRAWB~1\\perl\\site\\lib',
          'sitelib' => 'C:\\STRAWB~1\\perl\\site\\lib',
          'sitelibexp' => 'C:\\STRAWB~1\\perl\\site\\lib',
          'vendorarch' => 'C:\\STRAWB~1\\perl\\vendor\\lib',
          'vendorarchexp' => 'C:\\STRAWB~1\\perl\\vendor\\lib',
          'vendorlib' => 'C:\\STRAWB~1\\perl\\vendor\\lib',
          'vendorlibexp' => 'C:\\STRAWB~1\\perl\\vendor\\lib'

Instead of trying to wade through it with File::Spec, just turn short names
into longnames and every \ into / and hope it works.

Also make it so that the MD5 line will align correctly on an 80-col terminal
(on wider terminals it won't matter either way), and show *all* the hashes
at all times (omitting the information makes little sense).

In addition some clarification to the undef/empty ->VERSION return error
and the diag contents themselves (Mithaldu++)