Clarify name of guard function from 3d56e026, add it globally
Instead of just marking all my DESTROYs with the function, just add it to a
base class, thus capturing *everything* there is to find. Yes it will be a
tad slower. And yes it will have a massive benefit to any DBIC user happening
to be caught in an unfriendly app-space