(optdeps) Fix incorrect optdep output introduced in 34d2deae / 31c31b8d
Too many thinkos for a simple feature like this. When given "foo" cpanm
actually tries to find '"foo"' on CPAN, and nothing works. The original
thinking was that the > in ~>= needs escaping when copy/pasted by a user, so
the ""-enclosing seemed like a valid approach. Of course this does not work
at all when the output is piped into cpanm directly:

 perl -Ilib -MDBIx::Class::Optional::Dependencies=-list_missing,deploy | cpanm

Moreover, all of the above is *utterly unnecessary*, as Foo~>=2 is entirely
equivalent to Foo~2, which requires no escaping (thus quoting) whatsoever.

Hopefully this is the end of the embarrassing flailing on this feature :(