(travis) Peg Module::Install - new versions deps on autodie, which is 5.8.4+
The actual problem is actually couple-notches-older than the peg - the autodie
use got introduced in tests at 1.14 (Oct 2014), but I never noticed it as I
install M::I with notest.

While autodie might be fixable, it's a distraction at this moment, and besides
the M::I builder is about to be retired anyway. Punting.