(travis) Add Schmorp's "stable perl" fork to the smoke rosters
A note on motivation: I personally fully support Marc Lehmann's assertion
that the current direction of the perl5 development effort leaves a lot to
be desired (put in the mildest terms possible). The amount of changes[1]
necessary to get a *very* widely used module working again, and the failure
of the original authors to abandon the problematic direction *on their own*
is nothing short of an embarrassment. I fully recognize the reason why Marc
elected to attempt to maintain a fork (a nontrivial long-term commitment),
instead of engaging in another round of politicking and downright begging.

The inclusion of "stability perl" in the DBIC full-test matrix is the very
least I can do to support this effort.
                                                            --ribasushi

[1] http://stableperl.schmorp.de/dist/stableperl-5.22.0-1.001.diff